[Feature]#137: Add pandas DF -> AnnData and AnnData -> pandas DF
- implemented a function to convert a given Anndata object to a pandas dataframe

- moved the existing df_to_anndata function from IO part to util part and expose it via the API
